refactor: address comprehensive code review feedback
Major improvements:
- Extract bookmark functionality to separate module for better maintainability
- Change CSS class from 'notion-bookmark' to 'bookmark' for better semantics
- Remove unnecessary trim_end() calls from test expectations
- Simplify import structure and remove unused imports

File organization:
- Create bookmark.rs module with all OGP and rich bookmark functionality
- Reduce converter.rs from 970 to ~540 lines
- Maintain clear separation of concerns between modules

Code quality improvements:
- Follow Rust best practices for module organization
- Maintain comprehensive test coverage (57 tests passing)
- Preserve all existing functionality while improving structure
